+++ packages/router/src/bindings/contractReader/index.ts
@@ -49,9 +49,14 @@ export const bindContractReader = async () => {
     try {
       transactions = await contractReader.getActiveTransactions();
       if (transactions.length > 0) {
-        logger.info("Got active transactions", requestContext, methodContext, { transactions: transactions.length });
+        logger.info("Got active transactions", requestContext, methodContext, {
+          transactions: transactions.length,
+        });
         logger.debug("Got active transactions", requestContext, methodContext, {
-          transactions: transactions,
+          transactions,
+        });
+        logger.debug("handling tracker", requestContext, methodContext, {
+          handlingTrackerLength: handlingTracker.size,
           handlingTracker: [...handlingTracker],
         });
       }
@@ -114,16 +119,24 @@ export const handleActiveTransactions = async (transactions: ActiveTransaction<a
       blockNumber: -1,
       chainId,
     });
-    const res = await handleSingle(transaction, requestContext);
-    logger.debug("Transaction Result", requestContext, methodContext, { transactionResult: res });
-    if (res) {
-      handlingTracker.set(transaction.crosschainTx.invariant.transactionId, {
-        blockNumber: res.blockNumber,
-        chainId,
+
+    handleSingle(transaction, requestContext)
+      .then((result) => {
+        logger.debug("Handle Single Result", requestContext, methodContext, { transactionResult: result });
+
+        if (result && result.blockNumber) {
+          handlingTracker.set(transaction.crosschainTx.invariant.transactionId, {
+            blockNumber: result.blockNumber,
+            chainId,
+          });
+        } else {
+          handlingTracker.delete(transaction.crosschainTx.invariant.transactionId);
+        }
+      })
+      .catch((err) => {
+        logger.debug("Handle Single Errors", requestContext, methodContext, { error: jsonifyError(err) });
+        handlingTracker.delete(transaction.crosschainTx.invariant.transactionId);
       });
-    } else {
-      handlingTracker.delete(transaction.crosschainTx.invariant.transactionId);
-    }
     await delay(750); // delay here to not flood the provider
   }
 };