+++ packages/sdk/src/sdk.ts
@@ -166,7 +166,7 @@ export class NxtpSdk {
     } else {
       let _natsUrl;
       let _authUrl;
-      switch (network) {
+      switch (this.config.network) {
         case "mainnet": {
           _natsUrl = natsUrl ?? (isNode() ? NATS_CLUSTER_URL : NATS_WS_URL);
           _authUrl = authUrl ?? NATS_AUTH_URL;
@@ -244,7 +244,12 @@ export class NxtpSdk {
       txManagerConfig,
       this.logger.child({ module: "TransactionManager" }, "debug"),
     );
-    this.subgraph = new Subgraph(signer, subgraphConfig, this.logger.child({ module: "Subgraph" }), skipPolling);
+    this.subgraph = new Subgraph(
+      signer,
+      subgraphConfig,
+      this.logger.child({ module: "Subgraph" }),
+      this.config.skipPolling,
+    );
   }
 
   async connectMessaging(bearerToken?: string): Promise<string> {
@@ -547,22 +552,22 @@ export class NxtpSdk {
             getDecimals(receivingAssetId, receivingProvider),
           ]);
 
-          const amtMinusGas = BigNumber.from(data.bid.amountReceived).sub(data.gasFeeInReceivingToken);
-          const lowerBound = calculateExchangeWad(
-            BigNumber.from(amtMinusGas),
+          let lowerBound = calculateExchangeWad(
+            BigNumber.from(amount),
             inputDecimals,
             lowerBoundExchangeRate,
             outputDecimals,
           );
+
+          lowerBound = lowerBound.sub(data.gasFeeInReceivingToken);
 
           // safe calculation if the amountReceived is greater than 4 decimals
           if (BigNumber.from(data.bid.amountReceived).lt(lowerBound)) {
             const msg = "Invalid bid price: price impact is more than the slippage tolerance";
             this.logger.warn(msg, requestContext, methodContext, {
               signer,
-              lowerBound: lowerBound,
+              lowerBound: lowerBound.toString(),
               bidAmount: data.bid.amount,
-              amtMinusGas: amtMinusGas.toString(),
               gasFeeInReceivingToken: data.gasFeeInReceivingToken,
               amountReceived: data.bid.amountReceived,
               slippageTolerance: slippageTolerance,