+++ packages/txservice/src/types.ts
@@ -1,7 +1,7 @@
 import { Logger, RequestContext } from "@connext/nxtp-utils";
 import { BigNumber, BigNumberish, providers, utils } from "ethers";
 
-import { MaxBufferLengthError } from "./error";
+import { MaxBufferLengthError, TransactionBackfilled } from "./error";
 
 export type ReadTransaction = {
   chainId: number;
@@ -116,15 +116,16 @@ type TransactionConfig = {
 type LoggableTransactionData = {
   txsId: string;
   nonce: number;
-  hashes: string[];
-  attempt: number;
+  hashes?: string[];
+  attempt?: number;
+  bumps?: number;
+  minedBlockNumber?: number;
+  confirmedBlockNumber?: number;
   gasPrice: string;
   gasLimit: string;
-  discontinued: boolean;
   error: any;
-  confirmations: number;
-  didSubmit: boolean;
-  didFinish: boolean;
+  confirmations?: number;
+  state?: string;
 };
 
 /**
@@ -140,20 +141,33 @@ export class Transaction {
 
   // Receipt that we received for the on-chain transaction that was mined with
   // the desired number of confirmations.
-  public receipt?: providers.TransactionReceipt;
+  private _receipt: providers.TransactionReceipt | undefined;
+  public get receipt(): providers.TransactionReceipt | undefined {
+    return this._receipt;
+  }
+  public set receipt(value: providers.TransactionReceipt | undefined) {
+    // If the receipt has not yet been set, we can assume this is the mined receipt.
+    if (!this._receipt) {
+      this.minedBlockNumber = value?.blockNumber ?? -1;
+    }
+    this._receipt = value;
+  }
+  // Used to track when the transaction was mined, in the event that the mined receipt gets replaced by
+  // the confirmation receipt.
+  public minedBlockNumber = -1;
 
   // TODO: private setter
   // Timestamp initially set on creation, but will be updated each time submit() is called.
   public timestamp: number = Date.now();
-
-  // Indicates whether this transaction has been killed by monitor loop.
-  public discontinued = false;
 
   // This error will be set in the instance of a failure.
   public error: Error | undefined;
 
   // Which transaction attempt we are on.
   public attempt = 0;
+
+  // How many times we have bumped the gas price.
+  public bumps = 0;
 
   public get hash(): string | undefined {
     return this.didMine
@@ -206,17 +220,21 @@ export class Transaction {
 
   public get loggable(): LoggableTransactionData {
     return {
+      // These values should always be defined.
       txsId: this.uuid,
       nonce: this.nonce,
-      attempt: this.attempt,
-      hashes: this.hashes,
       gasPrice: `${utils.formatUnits(this.gas.price, "gwei")} gwei`,
       gasLimit: this.gas.limit.toString(),
-      discontinued: this.discontinued,
+      // Keeping these values as undefined if 0 in order to keep logs slim (undefined values won't be logged).
+      attempt: this.attempt > 0 ? this.attempt : undefined,
+      bumps: this.bumps > 0 ? this.bumps : undefined,
+      hashes: this.hashes.length > 0 ? this.hashes : undefined,
+      // Track block numbers for mine and confirm for observability.
+      minedBlockNumber: this.minedBlockNumber === -1 ? undefined : this.minedBlockNumber,
+      confirmedBlockNumber: this.receipt && (this.receipt.blockNumber > this.minedBlockNumber) ? this.receipt.blockNumber : undefined,
+      confirmations: this.receipt?.confirmations,
       error: this.error,
-      confirmations: this.receipt ? this.receipt.confirmations : 0,
-      didSubmit: this.didSubmit,
-      didFinish: this.didFinish,
+      state: this.error ? "E" : this.didFinish ? "C" : this.didMine ? "M" : this.didSubmit ? "S" : undefined,
     };
   }
 
@@ -252,10 +270,8 @@ export class TransactionBuffer extends Array<Transaction> {
     return this.map((tx) => tx.nonce);
   }
 
-  private get lastNonce(): number | undefined {
-    const nonces = this.nonces;
-    return nonces.length > 0 ? nonces[nonces.length - 1] : undefined;
-  }
+  // We use this to record the last tx that was shifted out of the buffer.
+  private lastShiftedTx: Transaction | undefined;
 
   /**
    * A data structure used for management of the lifecycle of on-chain transactions.
@@ -277,31 +293,71 @@ export class TransactionBuffer extends Array<Transaction> {
     super();
   }
 
+  // TODO: Check to make sure no async mutex issues could occur. If so, we should probably use a serialized queue for mutex.
+  /**
+   * Adds a transaction to the buffer.
+   * @param tx - The transaction to add to the buffer.
+   * @returns number index.
+   */
   public push(tx: Transaction): number {
-    const lastNonce = this.lastNonce;
-    if (lastNonce && tx.nonce < lastNonce) {
-      this.log(
-        "A nonce was pushed out of order. Please report this incident",
-        {
-          tx: tx.loggable,
-          lastNonce,
-        },
-        true,
-      );
-    }
-
     if (this.isFull) {
       throw new MaxBufferLengthError({
         maxLength: this.maxLength,
       });
     }
-    const index = super.push(tx);
-    this.log();
+
+    let index;
+    const nonces = this.nonces;
+    const lastNonce = nonces[nonces.length - 1] ?? this.lastShiftedTx?.nonce ?? -1;
+    const backfill = tx.nonce <= lastNonce;
+    if (!backfill) {
+      // If the nonce is greater than the last nonce, we can just push the tx to the end of the array.
+      index = super.push(tx);
+      if (lastNonce !== -1 && tx.nonce > lastNonce + 1) {
+        // Log if this nonce increments from the last by more than 1. This could be normal in the event that
+        // we're "catching up" to the correct nonce, thus having to skip a few nonces.
+        this.log("Pushed transaction skipped nonce value(s).", {
+          lastNonce,
+          gap: tx.nonce - lastNonce - 1,
+          transaction: tx.loggable,
+        });
+      }
+    } else {
+      // If the nonce is less than (or equal to) the last nonce, we need to find the correct index to insert the tx
+      // for a backfill. If the nonce does exist already, we'll overwrite that transaction and error it out.
+      let replacedTx;
+      if (this.lastShiftedTx && this.lastShiftedTx.nonce === tx.nonce) {
+        // If the lastShiftedTx (possibly being operated on by mine loop currently) is our replacement/backfill target,
+        // we should error that tx out.
+        super.unshift(tx);
+        index = 0;
+        replacedTx = this.lastShiftedTx;
+      } else {
+        // Loop through the array and find the correct index to insert the tx.
+        for (index = 0; index < this.length; index++) {
+          if (this[index].nonce >= tx.nonce) {
+            break;
+          }
+        }
+        // This splice operation will replace that transaction in the buffer.
+        replacedTx = super.splice(index, 1, tx)[0];
+      }
+      replacedTx.error = new TransactionBackfilled({
+        replacement: tx.loggable,
+      });
+    }
+
+    this.log(undefined, { index, nonce: tx.nonce, backfill });
     return index;
   }
 
+  /**
+   * Shifts a transaction from the buffer.
+   * @returns The last transaction in the buffer.
+   */
   public shift(): Transaction | undefined {
     const tx = super.shift();
+    this.lastShiftedTx = tx;
     this.log();
     return tx;
   }