+++ packages/txservice/test/provider.spec.ts
@@ -62,6 +62,7 @@ describe("ChainRpcProvider", () => {
       {
         ...DEFAULT_CONFIG,
         gasInitialBumpPercent: 20,
+        gasPriceMaxIncreaseScalar: 200,
       },
       signer,
     );
@@ -357,20 +358,37 @@ describe("ChainRpcProvider", () => {
 
     it("should bump gas price up to minimum if it is below that", async () => {
       // For test reliability, start from the config value and work backwards.
-      const expectedGas = (chainProvider as any).config.gasMinimum;
-      const testGasPrice = BigNumber.from(expectedGas)
+      const expectedGasPrice = (chainProvider as any).config.gasMinimum;
+      const testGasPrice = BigNumber.from(expectedGasPrice)
         .sub(
-          BigNumber.from(expectedGas)
+          BigNumber.from(expectedGasPrice)
             .mul((chainProvider as any).config.gasInitialBumpPercent)
             .div(100),
         )
-        .sub(utils.parseUnits("1", "gwei") as BigNumber);
+        .sub(utils.parseUnits("1", "gwei"));
       coreProvider.getGasPrice.resolves(testGasPrice);
 
       const result = await (chainProvider as any).getGasPrice();
 
       expect(result.isOk()).to.be.true;
-      expect(result.isOk() ? result.value.toString() : null).to.be.eq(expectedGas);
+      expect(result.isOk() ? result.value.toString() : null).to.be.eq(expectedGasPrice);
+    });
+
+    it("should employ the gas price max increase scalar if configured and applicable", async () => {
+      // For test reliability, start from the config value and work backwards.
+      const testScalar = (chainProvider as any).config.gasPriceMaxIncreaseScalar;
+      const testLastUsedGasPrice = utils.parseUnits("5", "gwei");
+      (chainProvider as any).lastUsedGasPrice = testLastUsedGasPrice;
+      // We're going to set the gas price our provider returns to the max value + 1 gwei.
+      // We expect the getGasPrice method to cap the price it returns at the max value.
+      const expectedGasPrice = testLastUsedGasPrice.mul(testScalar).div(100);
+      const testGasPrice = expectedGasPrice.add(utils.parseUnits("1", "gwei"));
+      coreProvider.getGasPrice.resolves(testGasPrice);
+
+      const result = await (chainProvider as any).getGasPrice();
+
+      expect(result.isOk()).to.be.true;
+      expect(result.isOk() ? result.value.toString() : null).to.be.eq(expectedGasPrice.toString());
     });
   });